requirements.txt
# TradingBot ML — Python Dependencies

# === Broker & Data ===
alpaca-py>=0.21.0          # Alpaca trading SDK
fredapi>=0.5.1             # FRED economic data
finnhub-python>=2.4.18     # Finnhub news/market data
alpha-vantage>=2.3.1       # Alpha Vantage financial data
yfinance>=0.2.31           # Yahoo Finance (backup data source)

# === Data Processing ===
pandas>=2.1.0
numpy>=1.25.0
ta>=0.11.0                 # Technical indicators (replaces pandas-ta)
scipy>=1.11.0

# === ML / Deep Learning ===
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.1.0
torch>=2.1.0               # PyTorch for LSTM, TFT, and TabNet
transformers>=4.35.0        # HuggingFace (for FinBERT)
# TFT (Temporal Fusion Transformer) and TabNet are implemented in pure PyTorch — no extra deps

# === NLP ===
nltk>=3.8.1
beautifulsoup4>=4.12.0     # Web scraping
requests>=2.31.0

# === Web Dashboard ===
flask>=3.0.0
flask-socketio>=5.3.0      # Real-time updates
plotly>=5.18.0              # Interactive charts
gunicorn>=21.2.0

# === Database ===
sqlalchemy>=2.0.0

# === Scheduling ===
apscheduler>=3.10.4

# === Backtesting ===
# (backtesting is built-in, no external library needed)

# === Utilities ===
pyyaml>=6.0.1
python-dotenv>=1.0.0
loguru>=0.7.2              # Better logging
tqdm>=4.66.0               # Progress bars
joblib>=1.3.0              # Model serialization
